requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
docker>=7.0.0
psycopg2-binary>=2.9.9
psutil>=5.9.8
pydantic>=2.0.0
